doc/cephadm: mention 'echo' being used during bootstrap w/ sudo
'cephadm bootstrap' runs "sudo echo" once to verify if the chosen
non-root user is allowed to run commands through sudo without password.
This is only required once on the node that does the bootstrapping and
a non-root user is used and can be removed from the list after the
bootstrap command has completed.